+++ maiacore/src/maiacore/python_wrapper/py_chord.cpp
@@ -324,9 +324,44 @@ void ChordClass(const py::module& m) {
     cls.def("getHarmonicSpectrum", &Chord::getHarmonicSpectrum, py::arg("numPartials") = 6,
             py::arg("amplCallback") = nullptr);
 
-    cls.def("getSetharesDissonanceValue", &Chord::getSetharesDissonanceValue,
-            py::arg("numPartials") = 6, py::arg("useMinModel") = true,
-            py::arg("amplCallback") = nullptr);
+    cls.def("getSetharesDissonance", &Chord::getSetharesDissonance, py::arg("numPartials") = 6,
+            py::arg("useMinModel") = true, py::arg("amplCallback") = nullptr,
+            py::arg("dissCallback") = nullptr);
+
+    cls.def(
+        "getSetharesPartialsDataFrame",
+        [](const Chord& chord, const int numPartials, const bool useMinModel,
+           const std::function<std::vector<float>(std::vector<float>)> amplCallback) {
+            const SetharesDissonanceTable table =
+                chord.getSetharesPartialsDissonanceValue(numPartials, useMinModel, amplCallback);
+
+            // Import Pandas module
+            py::object Pandas = py::module_::import("pandas");
+
+            // Get method 'from_records' from 'DataFrame()' object
+            py::object FromRecords = Pandas.attr("DataFrame").attr("from_records");
+
+            // Set DataFrame columns name
+            std::vector<std::string> columns = {"freqBaseIdx",   "freqBase",   "ampBase",
+                                                "freqTargetIdx", "freqTarget", "ampTarget",
+                                                "calcAmplitude", "dissonance"};
+
+            // Fill DataFrame with records and columns
+            py::object df = FromRecords(table, "columns"_a = columns);
+
+            df.attr("freqBaseIdx") = df.attr("freqBaseIdx").attr("astype")("int16");
+            df.attr("freqBase") = df.attr("freqBase").attr("astype")("float32");
+            df.attr("ampBase") = df.attr("ampBase").attr("astype")("float32");
+            df.attr("freqTargetIdx") = df.attr("freqTargetIdx").attr("astype")("int16");
+            df.attr("freqTarget") = df.attr("freqTarget").attr("astype")("float32");
+            df.attr("ampTarget") = df.attr("ampTarget").attr("astype")("float32");
+            df.attr("calcAmplitude") = df.attr("calcAmplitude").attr("astype")("float32");
+            df.attr("dissonance") = df.attr("dissonance").attr("astype")("float32");
+
+            return df;
+        },
+        py::arg("numPartials") = 6, py::arg("useMinModel") = true,
+        py::arg("amplCallback") = nullptr);
 
     cls.def(py::self == py::self);
     cls.def(py::self != py::self);